requirements.txt
pywhatkit==5.3
requests==2.28.0
python-dotenv==0.20.0
